// src/interfaces/Product_Commodity_Association.interface.ts
export interface ProductCommodityAssociation {
     product_commodity_association_id: number;
     commodity_id: number;
     product_id: number;
     product_category_types_id: number;
     measurement_unit: string;
     quantity: number;
     status: number;
     created_on: Date;
     created_by: number;
     updated_on: Date;
     updated_by: number;
  }
  